Feat: add separate class for game print

+++ srcs/en/swingy/equipement/helm/Helm.java
@@ -5,7 +5,7 @@ import en.swingy.equipement.Equipement;
 public class Helm extends Equipement {
 	protected final float hit_points;
 
-	public Helm(String p_name, int p_hit_points) {
+	public Helm(String p_name, float p_hit_points) {
 		super(p_name);
 		this.hit_points = p_hit_points;
 	}
