+++ srcs/en/swingy/equipement/weapon/Weapon.java
@@ -5,7 +5,7 @@ import en.swingy.equipement.Equipement;
 public class Weapon extends Equipement {
 	protected final float attack;
 
-	public Weapon(String p_name, int p_attack) {
+	public Weapon(String p_name, float p_attack) {
 		super(p_name);
 		this.attack = p_attack;
 	}
